fix for po and translation environment
