+++ po/.gitignore
@@ -1,2 +1,3 @@
-*
+template.pot
+*.mo
 !.gitignore
