+++ www/js/frontend/.gitignore
@@ -1,0 +1,1 @@
+translations.js
