+++ www/js/frontend/app.js
@@ -1,15 +1,15 @@
 define([
   'frontend/controllers/appController',
   'frontend/controllers/loginController',
-  'frontend/translations',
+  'frontend/translateBase',
   'angular',
   'angular-sanitize',
   'angular-gettext',
-  'onsen'], function (AppController, LoginController, Translations) {
+  'onsen'], function (AppController, LoginController, TranslateBase) {
 
 var app = angular.module('app', ['onsen', 'ngSanitize', 'gettext']);
 
-app.run(Translations);
+app.run(TranslateBase);
 app.controller('LoginController', LoginController);
 app.controller('AppController', AppController);
