+++ www/js/frontend/translations.js
@@ -1,18 +1,6 @@
-define(function(require, exports, module) {
-var translationTexts = {};
-translationTexts['en'] = require('translation/en');
-translationTexts['zh_TW'] = require('translation/zh_tw');
-
-// FIXME: swichable language support
-var lang = window.navigator.userLanguage || window.navigator.language || "";
-lang = lang.toLowerCase().match(/^zh/) ? 'zh_TW' : 'en';
-
-var translations = ['gettextCatalog', function (gettextCatalog) {
-  gettextCatalog.currentLanguage = lang;
-  gettextCatalog.debug = true;
-  gettextCatalog.setStrings(lang, translationTexts[lang]);
-}];
-
-return translations;
-
-});
+angular.module('app').run(['gettextCatalog', function (gettextCatalog) {
+/* jshint -W100 */
+    gettextCatalog.setStrings('en', {"Close":"Close","Login":"Login"});
+    gettextCatalog.setStrings('zh_TW', {"Close":"關閉","Login":"登入"});
+/* jshint +W100 */
+}]);
