--- www/js/translation/en.js
@@ -1,10 +1,0 @@
-define([], function () {
-
-var en = {
-  'Login': 'Login',
-  'Close': 'Close'
-};
-
-return en;
-
-});
