--- www/js/translation/zh_tw.js
@@ -1,10 +1,0 @@
-define([], function () {
-
-var tw = {
-  'Login': '登入',
-  'Close': '關閉'
-};
-
-return tw;
-
-});
